Only get predicates when initMatcher
This seems like a micro performance optimization - getting predicates is really fast. And that is true. But, by only getting them at startup, it makes it reasonable to FAIL if there is a problem with match hooks. This will make it much easier to see if you made a mistake in your match hooks JSON